cleanup and first attempt at filtering in the frontend
